+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/registry/PolarisServiceRegistry.java
@@ -100,6 +100,7 @@ public class PolarisServiceRegistry implements ServiceRegistry<PolarisRegistrati
 			LOGGER.warn("No service to register for polaris client...");
 			return;
 		}
+		registration.customize();
 		String serviceId = registration.getServiceId();
 
 		// Register instance.