+++ spring-cloud-starter-tencent-polaris-ratelimit/src/main/java/com/tencent/cloud/polaris/ratelimit/constant/RateLimitConstant.java
@@ -18,34 +18,24 @@
 
 package com.tencent.cloud.polaris.ratelimit.constant;
 
-import org.springframework.core.Ordered;
-
 /**
  * Constant for rate-limiter.
  *
  * @author Haotian Zhang
  */
 public final class RateLimitConstant {
-
-
-	/**
-	 * Default Private Constructor .
-	 */
-	private RateLimitConstant() {
-	}
-
-	/**
-	 * Order of filter.
-	 */
-	public static final int FILTER_ORDER = Ordered.HIGHEST_PRECEDENCE + 10;
-
 	/**
 	 * Info of rate limit.
 	 */
 	public static String QUOTA_LIMITED_INFO = "The request is denied by rate limit because the throttling threshold is reached";
-
 	/**
 	 * The build in label method.
 	 */
 	public static String LABEL_METHOD = "method";
+
+	/**
+	 * Default Private Constructor.
+	 */
+	private RateLimitConstant() {
+	}
 }