+++ spring-cloud-starter-tencent-polaris-ratelimit/src/main/java/com/tencent/cloud/polaris/ratelimit/filter/QuotaCheckReactiveFilter.java
@@ -26,9 +26,9 @@ import java.util.Set;
 import javax.annotation.PostConstruct;
 
 import com.tencent.cloud.common.constant.HeaderConstant;
+import com.tencent.cloud.common.constant.OrderConstant;
 import com.tencent.cloud.common.metadata.MetadataContext;
 import com.tencent.cloud.polaris.ratelimit.config.PolarisRateLimitProperties;
-import com.tencent.cloud.polaris.ratelimit.constant.RateLimitConstant;
 import com.tencent.cloud.polaris.ratelimit.resolver.RateLimitRuleArgumentReactiveResolver;
 import com.tencent.cloud.polaris.ratelimit.spi.PolarisRateLimiterLimitedFallback;
 import com.tencent.cloud.polaris.ratelimit.utils.QuotaCheckUtils;
@@ -72,9 +72,9 @@ public class QuotaCheckReactiveFilter implements WebFilter, Ordered {
 	private String rejectTips;
 
 	public QuotaCheckReactiveFilter(LimitAPI limitAPI,
-									PolarisRateLimitProperties polarisRateLimitProperties,
-									RateLimitRuleArgumentReactiveResolver rateLimitRuleArgumentResolver,
-									@Nullable PolarisRateLimiterLimitedFallback polarisRateLimiterLimitedFallback) {
+			PolarisRateLimitProperties polarisRateLimitProperties,
+			RateLimitRuleArgumentReactiveResolver rateLimitRuleArgumentResolver,
+			@Nullable PolarisRateLimiterLimitedFallback polarisRateLimiterLimitedFallback) {
 		this.limitAPI = limitAPI;
 		this.polarisRateLimitProperties = polarisRateLimitProperties;
 		this.rateLimitRuleArgumentResolver = rateLimitRuleArgumentResolver;
@@ -88,7 +88,7 @@ public class QuotaCheckReactiveFilter implements WebFilter, Ordered {
 
 	@Override
 	public int getOrder() {
-		return RateLimitConstant.FILTER_ORDER;
+		return OrderConstant.Server.Reactive.RATE_LIMIT_FILTER_ORDER;
 	}
 
 	@Override
@@ -110,7 +110,8 @@ public class QuotaCheckReactiveFilter implements WebFilter, Ordered {
 					response.setRawStatusCode(polarisRateLimiterLimitedFallback.rejectHttpCode());
 					response.getHeaders().setContentType(polarisRateLimiterLimitedFallback.mediaType());
 					dataBuffer = response.bufferFactory().allocateBuffer()
-							.write(polarisRateLimiterLimitedFallback.rejectTips().getBytes(polarisRateLimiterLimitedFallback.charset()));
+							.write(polarisRateLimiterLimitedFallback.rejectTips()
+									.getBytes(polarisRateLimiterLimitedFallback.charset()));
 				}
 				else {
 					response.setRawStatusCode(polarisRateLimitProperties.getRejectHttpCode());
@@ -118,9 +119,12 @@ public class QuotaCheckReactiveFilter implements WebFilter, Ordered {
 					dataBuffer = response.bufferFactory().allocateBuffer()
 							.write(rejectTips.getBytes(StandardCharsets.UTF_8));
 				}
-				response.getHeaders().add(HeaderConstant.INTERNAL_CALLEE_RET_STATUS, RetStatus.RetFlowControl.getDesc());
+				response.getHeaders()
+						.add(HeaderConstant.INTERNAL_CALLEE_RET_STATUS, RetStatus.RetFlowControl.getDesc());
 				if (Objects.nonNull(quotaResponse.getActiveRule())) {
-					response.getHeaders().add(HeaderConstant.INTERNAL_ACTIVE_RULE_NAME, quotaResponse.getActiveRule().getName().getValue());
+					response.getHeaders()
+							.add(HeaderConstant.INTERNAL_ACTIVE_RULE_NAME, quotaResponse.getActiveRule().getName()
+									.getValue());
 				}
 				return response.writeWith(Mono.just(dataBuffer));
 			}