+++ spring-cloud-tencent-polaris-loadbalancer/src/main/java/com/tencent/cloud/polaris/loadbalancer/PolarisWeightedRandomRule.java
@@ -18,71 +18,24 @@
 
 package com.tencent.cloud.polaris.loadbalancer;
 
-import java.util.List;
-
-import com.netflix.client.config.IClientConfig;
-import com.netflix.loadbalancer.AbstractServerPredicate;
-import com.netflix.loadbalancer.AvailabilityPredicate;
-import com.netflix.loadbalancer.CompositePredicate;
-import com.netflix.loadbalancer.PredicateBasedRule;
-import com.netflix.loadbalancer.Server;
-import com.tencent.cloud.common.pojo.PolarisServer;
 import com.tencent.polaris.api.config.consumer.LoadBalanceConfig;
-import com.tencent.polaris.api.pojo.Instance;
-import com.tencent.polaris.api.pojo.ServiceInstances;
 import com.tencent.polaris.router.api.core.RouterAPI;
 import com.tencent.polaris.router.api.rpc.ProcessLoadBalanceRequest;
-import com.tencent.polaris.router.api.rpc.ProcessLoadBalanceResponse;
-
-import org.springframework.util.CollectionUtils;
 
 /**
  * Polaris weighted load balancer.
  *
  * @author lepdou 2022-05-17
  */
-public class PolarisWeightedRandomRule extends PredicateBasedRule {
-
-	private final RouterAPI routerAPI;
-	private CompositePredicate compositePredicate;
+public class PolarisWeightedRandomRule extends AbstractPolarisRule {
 
 	public PolarisWeightedRandomRule(RouterAPI routerAPI) {
-		this.routerAPI = routerAPI;
+		super(routerAPI);
 	}
 
 	@Override
-	public void initWithNiwsConfig(IClientConfig clientConfig) {
-		AvailabilityPredicate availabilityPredicate = new AvailabilityPredicate(this, clientConfig);
-		compositePredicate = CompositePredicate.withPredicates(availabilityPredicate).build();
-	}
-
-	@Override
-	public AbstractServerPredicate getPredicate() {
-		return compositePredicate;
-	}
-
-	@Override
-	public Server choose(Object key) {
-		List<Server> servers = getLoadBalancer().getReachableServers();
-		if (CollectionUtils.isEmpty(servers)) {
-			return null;
-		}
-
-		// filter circuit breaker servers by ribbon
-		if (compositePredicate != null) {
-			servers = compositePredicate.getEligibleServers(servers);
-		}
-
-		ServiceInstances serviceInstances = LoadBalancerUtils.transferServersToServiceInstances(servers);
-
-		ProcessLoadBalanceRequest request = new ProcessLoadBalanceRequest();
-		request.setDstInstances(serviceInstances);
+	protected ProcessLoadBalanceRequest setProcessLoadBalanceRequest(ProcessLoadBalanceRequest request) {
 		request.setLbPolicy(LoadBalanceConfig.LOAD_BALANCE_WEIGHTED_RANDOM);
-
-		ProcessLoadBalanceResponse processLoadBalanceResponse = routerAPI.processLoadBalance(request);
-
-		Instance targetInstance = processLoadBalanceResponse.getTargetInstance();
-
-		return new PolarisServer(serviceInstances, targetInstance);
+		return request;
 	}
 }